Separated deep dive into modules. Added richer visualization using ipycytoscape to provide graph visualizations for query trees. Enabled both simple older visualization and new visualization, by user choice.
